refactor: extract internal/tools
Move the tool registration plan/diff/naming logic out of
internal/taskfileserver into a dedicated internal/tools package as part
of the package split tracked in #80.

What moved:
- registeredTool, createToolForTask, toolPlan, buildToolPlan, toolsEqual,
  diffTools (formerly internal/taskfileserver/tools.go) become
  RegisteredTool, CreateToolForTask, Plan, BuildPlan, Equal, Diff in
  internal/tools/{plan,create,diff}.go.
- sanitizeToolName, isWildcardTask, countWildcards, sanitizeRootPrefix,
  rootPrefix, prefixedToolName, shortToolNameHash move to
  internal/tools/naming.go (exported where the planner crosses package
  boundaries; helpers stay unexported).
- toolStateSnapshot and toolRootSnapshot become tools.StateSnapshot and
  tools.RootSnapshot. Server.snapshotToolStateLocked still constructs
  them under lock.

What stayed:
- toolRegistry interface and Server.syncTools live in taskfileserver.
  syncTools is now a thin coordinator: snapshot under lock, build plan +
  diff via tools.BuildPlan / tools.Diff with no lock held, validate the
  generation, then apply via toolRegistry — preserving the
  snapshot/plan/apply pattern with the generation guard intact.

Tests follow source: planner-only tests (CreateToolForTask, BuildPlan,
Equal/Diff, Sanitize*) move to internal/tools; the syncTools-via-Server
integration tests and snapshot-isolation test stay in taskfileserver.

No behaviour change.
