+++ internal/taskfileserver/integration_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/modelcontextprotocol/go-sdk/jsonrpc"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 	"github.com/rsclarke/mcp-taskfile-server/internal/roots"
+	"github.com/rsclarke/mcp-taskfile-server/internal/tools"
 )
 
 func TestHandleInitialized_WithRoots(t *testing.T) {
@@ -32,7 +33,7 @@ func TestHandleInitialized_WithRoots(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -99,7 +100,7 @@ func TestHandleInitialized_CallToolExecutesSingleRootTool(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -155,7 +156,7 @@ func TestHandleInitialized_DeduplicatesEquivalentRootURIs(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -216,7 +217,7 @@ func TestHandleRootsChanged_AddAndRemove(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -282,7 +283,7 @@ func TestHandleRootsChanged_EquivalentURIAliasKeepsSingleRoot(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -368,7 +369,7 @@ func TestHandleInitialized_NoPublicTasks(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -422,7 +423,7 @@ func TestHandleRootsChanged_TransitionToUnprefixed(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -490,7 +491,7 @@ func TestHandleRootsChanged_TransitionToUnprefixedCallTool(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -552,7 +553,7 @@ func TestHandleRootsChanged_RemoveLastRootClearsTools(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
@@ -614,7 +615,7 @@ func TestToolListChangedNotification_OnFileChange(t *testing.T) {
 		RootsListChangedHandler: ts.HandleRootsChanged,
 	})
 	ts.toolRegistry = server
-	ts.registeredTools = make(map[string]registeredTool)
+	ts.registeredTools = make(map[string]tools.RegisteredTool)
 
 	ct, st := mcp.NewInMemoryTransports()
 	ss, err := server.Connect(ctx, st, nil)
