+++ internal/taskfileserver/state.go
@@ -5,9 +5,9 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/go-task/task/v3/taskfile/ast"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 	"github.com/rsclarke/mcp-taskfile-server/internal/roots"
+	"github.com/rsclarke/mcp-taskfile-server/internal/tools"
 )
 
 // toolRegistry is the subset of *mcp.Server used for tool registration.
@@ -31,7 +31,7 @@ type toolRegistry interface {
 type Server struct {
 	roots           map[string]*roots.Root
 	toolRegistry    toolRegistry
-	registeredTools map[string]registeredTool
+	registeredTools map[string]tools.RegisteredTool
 	watchers        *watcherManager
 	logger          atomic.Pointer[slog.Logger]
 	mu              sync.Mutex
@@ -45,34 +45,17 @@ func (s *Server) log() *slog.Logger {
 	return s.logger.Load()
 }
 
-// toolRootSnapshot is an immutable per-root view captured under lock so
-// the planner can run without holding s.mu and without dereferencing a
-// live *Root that another goroutine may mutate. The taskfile pointer is
-// captured by value: even if reloadRoot later swaps root.Taskfile, this
-// snapshot keeps observing the AST that was current at snapshot time.
-type toolRootSnapshot struct {
-	workdir  string
-	taskfile *ast.Taskfile
-}
-
-// toolStateSnapshot captures the inputs needed by buildToolPlan,
-// frozen at a specific generation.
-type toolStateSnapshot struct {
-	generation uint64
-	roots      map[string]toolRootSnapshot
-}
-
 // snapshotToolStateLocked returns a snapshot of the current tool-relevant
 // server state. The caller must hold s.mu.
-func (s *Server) snapshotToolStateLocked() toolStateSnapshot {
-	snap := toolStateSnapshot{
-		generation: s.generation,
-		roots:      make(map[string]toolRootSnapshot, len(s.roots)),
+func (s *Server) snapshotToolStateLocked() tools.StateSnapshot {
+	snap := tools.StateSnapshot{
+		Generation: s.generation,
+		Roots:      make(map[string]tools.RootSnapshot, len(s.roots)),
 	}
 	for uri, root := range s.roots {
-		snap.roots[uri] = toolRootSnapshot{
-			workdir:  root.Workdir,
-			taskfile: root.Taskfile,
+		snap.Roots[uri] = tools.RootSnapshot{
+			Workdir:  root.Workdir,
+			Taskfile: root.Taskfile,
 		}
 	}
 	return snap
